requirements.txt
Flask==3.1.1
psycopg2-binary==2.9.10
Flask-SQLAlchemy==3.1.1
